requirements.txt
absl-py==1.0.0
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.4.1
astunparse==1.6.3
attrs==21.2.0
autopep8==1.5.7
backcall==0.2.0
backports.entry-points-selectable==1.1.0
bleach==4.1.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.9
colorama==0.4.4
cycler==0.10.0
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
distlib==0.3.3
Django==4.0.2
django-classy-tags==0.6.2
django-clearcache==1.2.1
django-crispy-forms==1.13.0
django-embed-video==1.4.1
django-filter==21.1
django-youtube-thumbnail==1.0.1
entrypoints==0.3
filelock==3.3.2
flatbuffers==2.0
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.42.0
gunicorn==20.1.0
h5py==3.6.0
idna==3.3
imageio==2.10.4
importlib-metadata==4.8.2
ipykernel==6.2.0
ipython==7.26.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.1
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.2
mistune==0.8.4
MouseInfo==0.1.3
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.3
notebook==6.4.3
notebook-as-pdf==0.5.0
numpy==1.21.2
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.0
pafy==0.3.82
pandas==1.3.2
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycodestyle==2.7.0
pycparser==2.20
pyee==7.0.4
PyGetWindow==0.0.9
Pygments==2.10.0
PyMsgBox==1.0.9
pyparsing==2.4.7
PyPDF2==1.27.7
pyperclip==1.8.2
pyppeteer==0.2.2
PyRect==0.1.4
pyrsistent==0.18.0
PyScreeze==0.1.28
python-dateutil==2.8.2
pytweening==1.0.4
pytz==2021.1
PyWavelets==1.2.0
pywin32==301
pywinpty==1.1.3
pyzmq==22.2.1
qtconsole==5.1.1
QtPy==1.10.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scikit-image==0.18.3
scikit-learn==1.0.1
scipy==1.7.1
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.0
termcolor==1.1.0
terminado==0.11.1
testpath==0.5.0
threadpoolctl==2.2.0
tifffile==2021.11.2
toml==0.10.2
tornado==6.1
tqdm==4.62.3
traitlets==5.0.5
typing_extensions==4.0.1
tzdata==2021.5
urllib3==1.26.7
virtualenv==20.10.0
virtualenvwrapper-win==1.2.6
wcwidth==0.2.5
webencodings==0.5.1
websockets==8.1
Werkzeug==2.0.2
whitenoise==6.2.0
widgetsnbextension==3.5.1
wrapt==1.13.3
xlrd==2.0.1
zipp==3.6.0
